单线程版。可手势

+++ main/ui.cpp
@@ -3,7 +3,7 @@
 #include "key.h"
 #include "led.h"
 
-uint8_t g_main_ui_case = 0;
+int16_t g_main_ui_case = 0;
 uint8_t g_bat_charge_flag = 0;
 
 void task_ui(void *pvParameters)
