+++ main/ui.h
@@ -8,7 +8,7 @@
  extern "C" {
 #endif
 
-extern uint8_t g_main_ui_case;
+extern int16_t g_main_ui_case;
 extern uint8_t g_bat_charge_flag;
 
 void task_ui(void *pvParameters);
